more i18n

+++ src/components/Account/Dashboard/AccountDashboard.jsx
@@ -3,22 +3,23 @@ import { useContext } from "react";
 import { AuthContext } from "../../AuthContext";
 import { Link } from "react-router-dom";
 import { useLocation } from "react-router-dom";
-import "./AccountDashboard.css"; // Dodajemy osobny plik CSS
+import { toast } from "react-toastify";
+import { FaTrash, FaSignOutAlt } from "react-icons/fa";
+import "./AccountDashboard.css";
+import "react-toastify/dist/ReactToastify.css";
 
 const AccountDashboard = () => {
-  const {
-    user,
-    logout,
-    deleteSnippet,
-    fetchAccountInfo,
-    changeNickname,
-  } = useContext(AuthContext);
+  const { user, logout, deleteSnippet, fetchAccountInfo, changeNickname } =
+    useContext(AuthContext);
   const location = useLocation();
   const [newNickname, setNewNickname] = useState("");
+  const [loading, setLoading] = useState(false);
 
   useEffect(() => {
-    if (user !== null) fetchAccountInfo(); // refresh
-    // eslint-disable-next-line react-hooks/exhaustive-deps
+    if (user !== null) {
+      setLoading(true);
+      fetchAccountInfo().finally(() => setLoading(false));
+    }
   }, [location]);
 
   const logoutHandler = () => {
@@ -27,51 +28,53 @@ const AccountDashboard = () => {
   };
 
   const handleDelete = async (id) => {
-    console.log("Deleting code with id:", id);
+    const confirmDelete = window.confirm(
+      "Are you sure you want to delete this snippet?"
+    );
+    if (!confirmDelete) return;
+
     try {
       await deleteSnippet(id);
+      toast.success("Snippet deleted successfully!");
     } catch (error) {
+      toast.error("Error deleting code");
       console.error("Error deleting code:", error);
     }
   };
 
-
-
   const handleNicknameChange = (e) => {
     e.preventDefault();
     changeNickname(newNickname);
-    setNewNickname("")
+    setNewNickname("");
+    toast.success("Nickname changed successfully!");
   };
 
   return (
     <>
-      {user && (
-        <div className="dashboard-container">
-          <h1 className="welcome-header">
-            Witaj, {user.userName || "Użytkowniku"}!
-          </h1>
-          {/* <p className="email-status">
-            Email: {user.email} potwierdzony:{" "}
-            {user.emailConfirmed ? "tak" : "nie"}
-          </p> */}
-
-          <button className="logout-button" onClick={logoutHandler}>
-            Wyloguj się
-          </button>
-          <h2>Twoje kody:</h2>
-          <div className="progress-bar">
-            <div
-              className="progress"
-              style={{ width: `${(user.codeSnippets.length / 10) * 100}%` }}
-            ></div>
-          </div>
-          <h3>
-            W darmowej wersji konta możesz posiadać maksymalnie 10 dokumentów
-          </h3>
-          {user && (
-            <ul className="code-list">
-              {user.codeSnippets.length > 0 ? (
-                user.codeSnippets.map((item) => (
+      {loading ? (
+        <div className="loading-spinner">Loading...</div>
+      ) : (
+        user && (
+          <div className="dashboard-container">
+            <h1 className="welcome-header">
+              Witaj, {user.userName || "Użytkowniku"}!
+            </h1>
+            <button className="logout-button" onClick={logoutHandler}>
+              <FaSignOutAlt /> Wyloguj się
+            </button>
+            <h2>Twoje kody:</h2>
+            <div className="progress-bar">
+              <div
+                className="progress"
+                style={{ width: `${(user.codeSnippets.length / 10) * 100}%` }}
+              ></div>
+            </div>
+            <h3>
+              W darmowej wersji konta możesz posiadać maksymalnie 10 dokumentów
+            </h3>
+            {user.codeSnippets.length > 0 ? (
+              <ul className="code-list">
+                {user.codeSnippets.map((item) => (
                   <li key={item.$id} className="code-item">
                     <Link to={"/" + item.uniqueId} className="code-link">
                       {item.uniqueId}
@@ -81,27 +84,27 @@ const AccountDashboard = () => {
                       className="delete-button"
                       onClick={() => handleDelete(item.uniqueId)}
                     >
-                      Usuń
+                      <FaTrash /> Usuń
                     </button>
                   </li>
-                ))
-              ) : (
-                <p>Jeszcze tu pusto</p>
-              )}
-            </ul>
-          )}
+                ))}
+              </ul>
+            ) : (
+              <p>Jeszcze tu pusto</p>
+            )}
 
-          <form className="nickname-form" onSubmit={handleNicknameChange}>
-            <label htmlFor="nickname">Nowy pseudonim:</label>
-            <input
-              type="text"
-              id="nickname"
-              value={newNickname}
-              onChange={(e) => setNewNickname(e.target.value)}
-            />
-            <button type="submit">Zmień pseudonim</button>
-          </form>
-        </div>
+            <form className="nickname-form" onSubmit={handleNicknameChange}>
+              <label htmlFor="nickname">Nowy pseudonim:</label>
+              <input
+                type="text"
+                id="nickname"
+                value={newNickname}
+                onChange={(e) => setNewNickname(e.target.value)}
+              />
+              <button type="submit">Zmień pseudonim</button>
+            </form>
+          </div>
+        )
       )}
     </>
   );
